fixed compilation with gcc4.3
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@740 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/cache_base.cpp
@@ -41,7 +41,7 @@ int CacheInterface::Read(std::string path, char* buf, size_t size, off_t off)
 		return 0;
 
 	/* Limit the read to the size of the file */
-	size_t to_read = off + size > file.GetFileSize() ? file.GetFileSize() - off : size;
+	size_t to_read = (size_t) ((off + size > file.GetFileSize()) ? file.GetFileSize() - off : size);
 
 	while(!file.HaveChunk(off, to_read))
 		usleep(10000);			  /* 0.01 sec */
